+++ apps/game_server_web/test/game_server/hooks/group_party_hooks_test.exs
@@ -173,7 +173,6 @@ defmodule GameServer.Hooks.GroupPartyHooksTest do
     def after_chat_message(_message), do: :ok
     @impl true
     def before_kv_get(_key, _opts), do: :public
-    @impl true
   end
 
   # ---------------------------------------------------------------------------